upgraded to react-native 0.31
