+++ Counter/app/containers/app.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react-native';
+import React, {Component} from 'react';
 import { createStore, applyMiddleware, combineReducers } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
