+++ Counter/app/containers/counterApp.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import React, { Component } from 'react-native';
+import React, {Component} from 'react';
 import {bindActionCreators} from 'redux';
 import Counter from '../components/counter';
 import * as counterActions from '../actions/counterActions';
